pyproject.toml
[tool.poetry]
name = "postgres_to_elastic"
version = "0.1.0"
description = ""
authors = ["Alex Vidyakov alex.vidyakov@yandex.ru"]

[tool.poetry.dependencies]
python = "3.9.6"
toml = "^0.10.2"
pydantic = "^1.8.2"
loguru = "^0.5.3"
psycopg2-binary = "^2.9.1"
requests = "^2.26.0"
backoff = "^1.11.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
